update scripts to only be run in correct directory

+++ CleaningScripts/clean_graphs.sh
@@ -1,4 +1,13 @@
 #!/bin/bash
 
-cd ../Graphs
-rm -rf *
+CURRENT_DIRECTORY=$(basename $PWD)
+
+if [ "${CURRENT_DIRECTORY}" == "CleaningScripts" ]; then
+    cd ../Graphs
+    echo "Cleaning graphs..."
+    rm -rf *
+    echo "Done."
+else
+    echo "Please execute script in CleaningScripts directory."
+    exit 1
+fi
