+++ CleaningScripts/clean_results.sh
@@ -1,4 +1,12 @@
 #!/bin/bash
 
-cd ../Code/Results
-rm -rf *
+CURRENT_DIRECTORY=$(basename $PWD)
+
+if [ "${CURRENT_DIRECTORY}" == "CleaningScripts" ]; then
+    cd ../Code/Results
+    echo "Cleaning results..."
+    rm -rf *
+    echo "Done."
+else
+    echo "Please execute script in CleaningScripts directory."
+fi
